Make the zlib dev package available in CentOS 8 and AmazonLinux2 images.
The intention of this change is to improve consistency between the various platform images. The libz headers and libraries are not a dependency of Swift at runtime on any OS, but are a common need for libraries such as `swift-nio-extras`, and are provided by the Ubuntu images. The resulting images were less than 60KB larger in local testing.