Add a new Dockerfile for building Swift-DocC-Render-Artifact
Adds a new Dockerfile that can be used to generate a built version of Swift-DocC-Render in isolation. This allows developers interested in having an up-to-date version of Swift-DocC-Render to build one in an automated way without installing Node.

Developers can build a copy of Swift-DocC-Render at the given branch with the included helper script by running something like the following:

    ./build-swift-docc-render.sh \
        --branch <branch-in-swift-docc-render-repo> \
        --output-path <path-to-output-directory>

The helper script starts by building a new version of the included Docker container with caching disabled. It then copies out the built artifact to the given output directory, and deletes the created Docker container.

Resolves r81634420.