Introduce a Dockerfile for Windows
This allows one to build a Windows container for Swift.  Although it is
not perfect, it is sufficient for basic tasks.

The Git installation should preferably enable symbolic link support
which although it is marked to do does not seem to be enabled.

The Python installation should preferably set `PYTHONHOME` and
`PYTHONPATH`.  This will become increasingly important as further
refinements to the toolchain start getting merged.