Symlink libstdc++_nonshared.a into static swift (#352)
The swift compiler and runtime libraries are built against the
devtoolset-8 on centOS-7. The devtoolset has additional functionality on
top of the base centos-7 libstdc++. When the standard library is built
as a dynamic library, the extra functionality is linked into the DSO
directly, so the dependency is hidden. When the standard library is
built as a static archive, it the extra functionality isn't. This is
correct behavior to avoid over-linking in the final executable, but it
means that the library needs to be present when building statically and
in a place that the compiler will look for resources.

This symlinks the backpack library with the additional functionality,
along with the linker script to tell the linker to link it into anything
that needs to use libstdc++.so.