Correct the padding on the /etc/motd banners (#379)
The prior tab based padding of the /etc/motd banner lended itself to an
error in the padding in length depending on the value of the Swift tag
contents. The approach of using `printf` is a bit more friendly from a
maintenance perspective.

Lastly, there was a bug in how the bash code was included where the
value of `$TERM` wasn't be escaped properly so you'd get a hardcoded
value vs. the variable `$TERM`.

Co-authored-by: Jon Miller <jonmiller@apple.com>