Skip cmake bootstrap
The cmake bootstrap is slower than using the installed CMake to build
the new CMake since it has to build a limited CMake first and
then use that for the full CMake.

This uses the existing CMake 3.22 from the container to build the CMake
from the Swift checkout with SSL disabled, testing disabled, and the
interactive curses UI disabled to improve the overall build times.