[Static Linux SDK] Build script improvements.
When running on an aarch64 host, we need to download a different version
of the Swift compiler in order to do the build.  If we run the build on
a macOS host, we also have to tell the update-checkout script to fetch
all the required dependencies (even Linux-only ones).

Also tell Foundation where swift-collections lives, to avoid it fetching
another copy.