[Static Linux SDK] Build script improvements. (#417)
* [Static Linux SDK] Build script improvements.

When running on an aarch64 host, we need to download a different version
of the Swift compiler in order to do the build.  If we run the build on
a macOS host, we also have to tell the update-checkout script to fetch
all the required dependencies (even Linux-only ones).

Also tell Foundation where swift-collections lives, to avoid it fetching
another copy.

* [CI] Run Docker builds in the right directory.

Make sure we run the Docker builds in the directory containing the
Dockerfile.  Also, prefer lists of arguments rather than strings,
especially when including filenames.