Configurable user IDs
When working with these containers to debug issues locally, it's
sometimes useful to have the source directory live on the host, while
building in the container. In order for permissions to line up, the user
ID numbers need to match or the build-user may not be able to create
the necessary directories.

This patch allows setting the build-user user ID so that the permissions
line up and the user in the container and host can manipulate the files.

Override the build user ID with the current user's ID is done with the
following command to ensure that the permissions on any volumes are
kept in sync is the following:
`docker build --build-arg BUILD_USER_ID=$(id -u) ...`