Use fixed version build tools link
The current link to the VS build tools points to an evergreen
bootstrapper. This means that when a new v17 buildtools.exe is
available the link is updated with the new version.

Since we're verifying that the downloaded buildtools.exe file hash
matches a known value the windows Docker setup breaks each time a new
buildtools is released.

Instead, use a fixed version permalink that will not change. The
specific version used is version `17.12.0`, build `17.12.35506.116`.

See: https://learn.microsoft.com/en-gb/visualstudio/releases/2022/release-history#fixed-version-bootstrappers