Install libc++ package on CI images of Ubuntu 24.04 (#420)
This adds a `libc++-18-dev` package to the list of packages installed on a CI image of Ubuntu 24.04.

Swift is now capable of compiling code that uses C++ interop with libc++ on Linux. Having the libc++ package installed provides us with CI coverage of the feature.

This is similar to https://github.com/swiftlang/swift-docker/pull/402.