[Static SDK] Build clang before running the main build. (#501)
* [Static SDK] Build clang before running the main build.

We need to do this because otherwise libc++ fails to build, since the
clang we pick up from the Swift install is an older version than the
one libc++ requires (libc++ currently needs Clang 19; we ship Clang
17 with Swift).

* [Static SDK] Turn off -Werror for BoringSSL.

`-Werror` is a stupid option; it basically breaks the build whenever
a newer compiler adds a new warning that triggers in your code.  We
don't need it here, so disable it.