[Static SDK] Tweak build script slightly.
We need to run the build and install of bzip2 together, otherwise
bzip2 tries to run tests, which won't necessarily work when cross
compiling for a different architecture.

We also want to remove various directories from the final result.

Oh, and we don't want to build libxml2 with lzma support either,
otherwise we'll have to link with liblxma whenever we use libxml2.